Moved recommended repos from ContainerStore to NewContainer.react.js
